requirements.txt
python-telegram-bot==21.6
google-generativeai==0.7.2
python-dotenv==1.0.1
httpx==0.27.0
